addressed comments and added AuthorisedDep dependency #550

+++ inventory_management_system_api/routers/v1/usage_status.py
@@ -13,6 +13,7 @@ from typing import Annotated
 
 from fastapi import APIRouter, Depends, HTTPException, Path, Request, status
 
+from inventory_management_system_api.auth.authorisation import AuthorisedDep
 from inventory_management_system_api.auth.jwt_bearer import JWTBearer
 from inventory_management_system_api.core.config import config
 from inventory_management_system_api.core.exceptions import (
@@ -31,7 +32,6 @@ router = APIRouter(prefix="/v1/usage-statuses", tags=["usage statuses"])
 
 UsageStatusServiceDep = Annotated[UsageStatusService, Depends(UsageStatusService)]
 
-
 @router.post(
     path="",
     summary="Create a new usage status",
@@ -39,18 +39,16 @@ UsageStatusServiceDep = Annotated[UsageStatusService, Depends(UsageStatusService
     status_code=status.HTTP_201_CREATED,
 )
 def create_usage_status(
-    usage_status: UsageStatusPostSchema, usage_status_service: UsageStatusServiceDep, request: Request
+    usage_status: UsageStatusPostSchema, usage_status_service: UsageStatusServiceDep, authorised: AuthorisedDep
 ) -> UsageStatusSchema:
     logger.info("Creating a new usage status")
     logger.debug("Usage status data: %s", usage_status)
 
     # check user is authorised to perform operation
-    if config.authentication.enabled is True:
-        jwt_bearer = JWTBearer()
-        if not jwt_bearer.is_jwt_access_token_authorised(request.state.token):
-            raise HTTPException(
-                status_code=status.HTTP_403_FORBIDDEN, detail="Not authorised to perform this operation"
-            )
+    if not authorised:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN, detail="Not authorised to perform this operation"
+        )
 
     try:
         usage_status = usage_status_service.create(usage_status)
@@ -101,16 +99,15 @@ def get_usage_status(
 def delete_usage_status(
     usage_status_id: Annotated[str, Path(description="ID of the usage status to delete")],
     usage_status_service: UsageStatusServiceDep,
-    request: Request,
+    authorised: AuthorisedDep
 ) -> None:
     logger.info("Deleting usage status with ID: %s", usage_status_id)
+    
     # check user is authorised to perform operation
-    if config.authentication.enabled is True:
-        jwt_bearer = JWTBearer()
-        if not jwt_bearer.is_jwt_access_token_authorised(request.state.token):
-            raise HTTPException(
-                status_code=status.HTTP_403_FORBIDDEN, detail="Not authorised to perform this operation"
-            )
+    if not authorised:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN, detail="Not authorised to perform this operation"
+        )
 
     try:
         usage_status_service.delete(usage_status_id)
